backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==1.4.49   # latest 1.4.x release
databases[postgresql]==0.8.0
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9   # ✅ required for Alembic & SQLAlchemy sync parts


# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ML and NLP
transformers==4.35.2
torch==2.2.2        # ✅ latest stable, works with Python 3.12
spacy==3.7.2
numpy==1.26.2       # ✅ matches Python 3.12
scikit-learn==1.3.2

# Geospatial
geopy==2.4.0
shapely==2.0.2
geoalchemy2==0.14.3
# Async and WebSocket
websockets==12.0
redis==5.0.1

# HTTP and API
httpx==0.25.2
requests==2.31.0

# Data processing
pandas==2.1.3
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
celery==5.3.4

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
